Safeguard deletion of `programmatic-types.graphql` in `artisan lighthouse:ide-helper`
https://github.com/nuwave/lighthouse/issues/1407

+++ src/Console/IdeHelperCommand.php
@@ -166,7 +166,7 @@ SDL;
 
         $filePath = static::programmaticTypesPath();
 
-        if ($types->isEmpty()) {
+        if ($types->isEmpty() && file_exists($filePath)) {
             unlink($filePath);
 
             return;
